requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.5
WTForms==3.0.1
SQLAlchemy==2.0.20
email-validator==2.0.0
WeasyPrint==59.0
Pillow==10.0.0
passlib==1.7.4
python-dotenv==1.0.0
flask-migrate==4.0.4
flask-babel==3.1.0
flask-admin==1.5.8
flask-uploads==0.2.1
markdown==3.4.4
pyjwt==2.8.0
requests==2.31.0
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
bootstrap-flask==2.2.0
flask-bootstrap==3.3.7.1
